% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabest.R
\name{stabest}
\alias{stabest}
\title{Estimation of student and college preferences in a m:1 school market. 
Identifying restrictions are imposed by (i) stability conditions of the observed matching, and 
(ii) the submitted rank order lists (which may be incomplete or contain unacceptable students / schools).}
\usage{
stabest(student.prefs, college.prefs, student.id = "s.id",
  college.id = "c.id", match.id = NULL, data, nSeats = NULL,
  demean = FALSE, niter = 500, thin = 10, burnin = 50,
  initparm = NULL, nCores = 1)
}
\arguments{
\item{student.prefs}{formula object, rank ~ explanatory variables. Encoding: lower rank = better, -1 = unacceptable / last, NA = unobserved. 
Careful when including factor variables with mutually exclusive levels - one level must be missing so as to not introduce an intercept through the back door!}

\item{college.prefs}{formula object, rank ~ explanatory variables. Encoding: lower rank = better, -1 = unacceptable, NA = unobserved
#' Careful when including factor variables with mutually exclusive levels - one level must be missing so as to not introduce an intercept through the back door!}

\item{student.id}{variable containing the IDs of the students}

\item{college.id}{variable containing the IDs of the schools}

\item{match.id}{variable indicating a match (1=match, 0=no match). If NULL, no stability conditions will be imposed.}

\item{data}{a data.frame}

\item{nSeats}{numeric vector indicating the number of seats for each school (order by college.id)}

\item{demean}{whether to de-mean variables by decision making unit (beta)}

\item{niter}{number of data augmentation iterations}

\item{thin}{thinning parameter}

\item{burnin}{number of discarded samples (in the thinned series, so burnin*thin must be less than niter)}

\item{initparm}{if not NULL, this must be a list with elements beta (coefficients driving students' prefs)
and gamma (coefficients driving schools' decisions). The number of coefficients must match the number of parameters
supplied via student.prefs and school.prefs, respectively (including factors).}

\item{nCores}{number of parallel cores to use (currently only 1 is permitted)}
}
\value{
An object of type "stabest"
}
\description{
WARNING: The method
}
\examples{

## Estimate students' and colleges' preferences with different identifying assumptions.

niter <- 100
thin <- 2
burnin <- 5
dat <- schoolmarket200
nSeats <- dat$sch_capacity[dat$stu_id==1]

## STABILTIY

fit1 <- stabest(.~-1 + distance + stu_score:sch_mscore + sch_FE, .~stu_score-1,
                data=dat, nSeats=nSeats, 
                student.id='stu_id',college.id='sch_id',match.id='sch_assignment',
                niter=niter, burnin=burnin, thin=thin 
)
summary(fit1)

## STABILITY + UNDOMINATED PREFERNCES

fit3 <- stabest(choice_rk~-1 + distance + stu_score:sch_mscore + sch_FE, 
               sch_rk_observed~stu_score-1,
                data=dat, nSeats=nSeats, 
                student.id='stu_id',college.id='sch_id',match.id='sch_assignment',
                niter=niter, burnin=burnin, thin=thin
)
summary(fit3)

## UNDOMINATED STRATEGIES

fit5 <- stabest(choice_rk~-1 + distance + stu_score:sch_mscore + sch_FE, 
                sch_rk_observed~stu_score-1,
                data=dat, nSeats=nSeats, 
                student.id='stu_id',college.id='sch_id',match.id=NULL,
                niter=niter, burnin=burnin, thin=thin
)
summary(fit5)

}
